- Teach the dragon to: speak and   breathe fire - Allow creation of the dragon from keyboard input -After all parameters are entered, the dragon should be output to the console -   use the toString() method

+++ src/lesson_16_classes/Homework.java
@@ -41,7 +41,7 @@ public class Homework {
         int age = scanner.nextInt();
 
         System.out.print("Введіть вагу: ");
-        double weight = scanner.nextInt();
+        double weight = scanner.nextDouble();
         scanner.nextLine();
 
         System.out.print("Введіть колір: ");
